+++ PocketPulse/iOS/Data/Services/NotificationService.swift
@@ -15,6 +15,7 @@ protocol NotificationServiceProtocol {
     func deleteNotification(id: UUID) async throws
 }
 
+
 @MainActor
 class NotificationService: NotificationServiceProtocol {
     private let context: ModelContext
@@ -24,36 +25,35 @@ class NotificationService: NotificationServiceProtocol {
     }
     
     func fetchNotifications() async throws -> [NotificationModel] {
-        let descriptor = FetchDescriptor<NotificationEntity>(sortBy: [SortDescriptor(\.date, order: .reverse)])
-        let entities = try context.fetch(descriptor)
-        return entities.map { $0.toModel() }
+        let descriptor = FetchDescriptor<NotificationModel>(sortBy: [SortDescriptor(\.date, order: .reverse)])
+        return try context.fetch(descriptor)
     }
     
     func markAllAsRead() async throws {
-        let descriptor = FetchDescriptor<NotificationEntity>() // Fetch all
-        let entities = try context.fetch(descriptor)
+        let descriptor = FetchDescriptor<NotificationModel>() // Fetch all
+        let notifications = try context.fetch(descriptor)
         
-        for entity in entities {
-            entity.isRead = true
+        for notification in notifications {
+            notification.isRead = true
         }
         
         try context.save()
     }
     
     func addNotification(title: String, message: String, type: NotificationCategory) async throws {
-        let entity = NotificationEntity(
+        let notification = NotificationModel(
             title: title,
             message: message,
             type: type
         )
-        context.insert(entity)
+        context.insert(notification)
         try context.save()
     }
     
     func deleteNotification(id: UUID) async throws {
-        let descriptor = FetchDescriptor<NotificationEntity>(predicate: #Predicate { $0.id == id })
-        if let entity = try context.fetch(descriptor).first {
-            context.delete(entity)
+        let descriptor = FetchDescriptor<NotificationModel>(predicate: #Predicate { $0.id == id })
+        if let notification = try context.fetch(descriptor).first {
+            context.delete(notification)
             try context.save()
         }
     }
